Player Movement developed
The player moves from side to side with impulses by pressing the spacebar to fly while using A or D to choose the direction.

The movement also takes in mind the air resistance with the "player.setDrag()" function.

The controls are implemented by adding a listener to the keyboard spacebar key so that the function is only called once per key press action.

The only thing left is to represent the movement bar and it's values and to adjust the increment of the barValue so that it isn't continuous but logarithmic instead.

Let's go!! 💪🔥